requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.7
django-environ==0.4.4
idna==2.6
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
